CMakeLists.txt: use correct path to script in source dir
The CMakeLists.txt file currently assumes that the source dir is one
level above the build dir, but this might not necessairily be the case.
Instead of ../, use the official CMAKE_BUILD_DIR variable that
references the directory where the CMakeLists.txt file is.

This avoids the following error when building, for example, with a cross
build system like ptxdist, that knows about how to build any kind of
CMake based project, and choses to put its build dir somewhere else:

  sh: 0: cannot open ../scripts/build/osDistro.sh: No such file

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>