CMakeLists.txt: Introduce pkg-config
For correct usage of header file inclusion and linking, it is more
correct to use pkg-config instead of ad-hoc methods. With pkg-config,
also special situations like out of tree building, setting a sysroot or
cross compiling can be handled properly.

The actual usage of pkg-config will follow in subsequent patches.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>